requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
Pillow==9.4.0
pypng==0.20220715.0
python-dateutil==1.5
qrcode==7.4.2
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
whitenoise==6.4.0
